Add files via upload
The game starts with the player in the woods, who then has to escape from the forest. The player encounters several characters throughout the game, such as, animals who need help, monsters who want to attack the player and etc. However, if the players choses to help the animals, the animals come back to save the player from the monsters. The players also encounters several herbs throughout the game which could be stored and used to increase the player's health.